backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
supabase==2.3.0
python-multipart==0.0.6
requests==2.32.3
mangum==0.17.0
openpyxl==3.1.2
cryptography==41.0.7